+++ Neolution.DotNet.Console.UnitTests/DotNetConsoleGrammarTests.cs
@@ -4,6 +4,7 @@
     using System.Threading.Tasks;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.DependencyInjection.Extensions;
+    using Neolution.DotNet.Console.Abstractions;
     using Neolution.DotNet.Console.UnitTests.Fakes;
     using Neolution.DotNet.Console.UnitTests.Spies;
     using Shouldly;
@@ -123,7 +124,7 @@
         /// <param name="args">The arguments.</param>
         /// <param name="tracker">The logger.</param>
         /// <returns>A built console app ready to run.</returns>
-        private static DotNetConsole CreateConsoleAppWithLogger(string args, IUnitTestLogger tracker)
+        private static IDotNetConsole CreateConsoleAppWithLogger(string args, IUnitTestLogger tracker)
         {
             var builder = DotNetConsole.CreateBuilderWithReference(Assembly.GetAssembly(typeof(DefaultCommand))!, args.Split(" "));