+++ Neolution.DotNet.Console/DotNetConsoleBuilder.cs
@@ -7,7 +7,6 @@
     using CommandLine;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.FileProviders;
     using Microsoft.Extensions.Hosting;
     using Microsoft.Extensions.Logging;
     using Neolution.DotNet.Console.Abstractions;
@@ -33,6 +32,11 @@
         /// The service collection
         /// </summary>
         private readonly ServiceCollection serviceCollection = new();
+
+        /// <summary>
+        /// Run only to verify dependencies.
+        /// </summary>
+        private bool verifyDependencies;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DotNetConsoleBuilder"/> class.
@@ -68,7 +72,7 @@
         /// Builds this instance.
         /// </summary>
         /// <returns>The <see cref="DotNetConsole"/>.</returns>
-        public DotNetConsole Build()
+        public IDotNetConsole Build()
         {
             // Copy over the services to host builder before it gets built
             this.hostBuilder.ConfigureServices(services =>
@@ -78,6 +82,13 @@
                     services.Add(service);
                 }
             });
+
+            if (this.verifyDependencies)
+            {
+                this.hostBuilder.UseEnvironment("Development");
+                this.hostBuilder.Build();
+                return new NoOperationConsole();
+            }
 
             var host = this.hostBuilder.Build();
             return new DotNetConsole(host, this.commandLineParserResult);
@@ -94,8 +105,14 @@
         /// </returns>
         internal static DotNetConsoleBuilder CreateBuilderInternal(Assembly assembly, Type[]? verbTypes, string[] args)
         {
+            // Create configuration and environment instances that are only valid before the host is built.
+            // We want to expose these as read-only properties in the DotNetConsoleBuilder.
+            var environment = DotNetConsoleDefaults.CreateConsoleEnvironment(args);
+            var configuration = DotNetConsoleDefaults.CreateConsoleConfiguration(assembly, args, environment);
+
             // Create a HostBuilder
             var builder = Host.CreateDefaultBuilder(args)
+                .UseContentRoot(environment.ContentRootPath)
                 .ConfigureLogging((context, logging) =>
                 {
                     AdjustDefaultBuilderLoggingProviders(logging);
@@ -109,19 +126,42 @@
                         .AsImplementedInterfaces());
                 });
 
-            // Manually build configuration and environment again, because we unfortunately can't access them from the host builder we just created, but want to provide them in the ConsoleApplicationBuilder.
-            var environment = CreateConsoleEnvironment(args);
-            var configuration = ApplyDefaultConfiguration(assembly, args, environment);
-
             // If verb types were not specified, compile all available verbs for this run by looking for classes with the Verb attribute in the specified assembly
             verbTypes ??= assembly.GetTypes()
                 .Where(t => t.GetCustomAttribute<VerbAttribute>() != null)
                 .ToArray();
 
-            EnforceStrictVerbMatching(args, verbTypes);
             var parsedArguments = Parser.Default.ParseArguments(args, verbTypes);
+            var consoleBuilder = new DotNetConsoleBuilder(builder, parsedArguments, environment, configuration);
 
-            return new DotNetConsoleBuilder(builder, parsedArguments, environment, configuration);
+            if (args.Length == 1 && string.Equals(args[0], "verify-dependencies", StringComparison.OrdinalIgnoreCase))
+            {
+                consoleBuilder.verifyDependencies = true;
+                return consoleBuilder;
+            }
+
+            CheckStrictVerbMatching(args, verbTypes);
+            return consoleBuilder;
+        }
+
+        /// <summary>
+        /// Adjusts the default builder logging providers.
+        /// </summary>
+        /// <param name="logging">The logging.</param>
+        private static void AdjustDefaultBuilderLoggingProviders(ILoggingBuilder logging)
+        {
+            // Remove the default logging providers
+            logging.ClearProviders();
+
+            // Re-add other logging providers that are assigned in Host.CreateDefaultBuilder
+            logging.AddDebug();
+            logging.AddEventSourceLogger();
+
+            if (OperatingSystem.IsWindows())
+            {
+                // Add the EventLogLoggerProvider on windows machines
+                logging.AddEventLog();
+            }
         }
 
         /// <summary>
@@ -130,7 +170,7 @@
         /// <param name="args">The arguments.</param>
         /// <param name="availableVerbTypes">The available verb types.</param>
         /// <exception cref="Neolution.DotNet.Console.DotNetConsoleException">Cannot create builder, because the specified verb '{firstVerb}' matches no command.</exception>
-        private static void EnforceStrictVerbMatching(string[] args, Type[] availableVerbTypes)
+        private static void CheckStrictVerbMatching(string[] args, Type[] availableVerbTypes)
         {
             var availableVerbs = availableVerbTypes.Select(t => t.GetCustomAttribute<VerbAttribute>()!).ToList();
             if (!availableVerbs.Any(v => v.IsDefault))
@@ -157,89 +197,6 @@
             if (!verbMatched)
             {
                 throw new DotNetConsoleException($"Cannot create builder, because the specified verb '{firstVerb}' matches no command.");
-            }
-        }
-
-        /// <summary>
-        /// Creates the console environment.
-        /// </summary>
-        /// <param name="args">The command line arguments.</param>
-        /// <returns>The <see cref="IHostEnvironment"/>.</returns>
-        private static DotNetConsoleEnvironment CreateConsoleEnvironment(string[] args)
-        {
-            var configuration = new ConfigurationBuilder()
-                .AddEnvironmentVariables(prefix: "DOTNET_")
-                .AddCommandLine(args)
-                .Build();
-
-            return new DotNetConsoleEnvironment
-            {
-                EnvironmentName = configuration[HostDefaults.EnvironmentKey] ?? "Production",
-                ApplicationName = AppDomain.CurrentDomain.FriendlyName,
-                ContentRootPath = AppContext.BaseDirectory,
-                ContentRootFileProvider = new PhysicalFileProvider(AppContext.BaseDirectory),
-            };
-        }
-
-        /// <summary>
-        /// Applies the default configuration to the console application.
-        /// </summary>
-        /// <param name="assembly">The assembly.</param>
-        /// <param name="args">The arguments.</param>
-        /// <param name="environment">The environment.</param>
-        /// <returns>The <see cref="IConfiguration" />.</returns>
-        private static IConfiguration ApplyDefaultConfiguration(Assembly assembly, string[] args, IHostEnvironment environment)
-        {
-            var configurationBuilder = new ConfigurationBuilder()
-                .SetBasePath(AppContext.BaseDirectory)
-                .AddEnvironmentVariables(prefix: "DOTNET_");
-
-            AddCommandLineConfig(configurationBuilder, args);
-
-            configurationBuilder
-                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                .AddJsonFile($"appsettings.{environment.EnvironmentName}.json", optional: true, reloadOnChange: true);
-
-            if (environment.IsDevelopment())
-            {
-                configurationBuilder.AddUserSecrets(assembly, optional: true, reloadOnChange: true);
-            }
-
-            configurationBuilder.AddEnvironmentVariables();
-
-            return configurationBuilder.Build();
-        }
-
-        /// <summary>
-        /// Adjusts the default builder logging providers.
-        /// </summary>
-        /// <param name="logging">The logging.</param>
-        private static void AdjustDefaultBuilderLoggingProviders(ILoggingBuilder logging)
-        {
-            // Remove the default logging providers
-            logging.ClearProviders();
-
-            // Re-add other logging providers that are assigned in Host.CreateDefaultBuilder
-            logging.AddDebug();
-            logging.AddEventSourceLogger();
-
-            if (OperatingSystem.IsWindows())
-            {
-                // Add the EventLogLoggerProvider on windows machines
-                logging.AddEventLog();
-            }
-        }
-
-        /// <summary>
-        /// Adds the command line configuration.
-        /// </summary>
-        /// <param name="configBuilder">The configuration builder.</param>
-        /// <param name="args">The arguments.</param>
-        private static void AddCommandLineConfig(IConfigurationBuilder configBuilder, string[]? args)
-        {
-            if (args is { Length: > 0 })
-            {
-                configBuilder.AddCommandLine(args);
             }
         }
     }